requirements.txt
telethon
httpx
python-dotenv
cryptg        # optional but speeds up downloads a lot